sbat: Set grub,4; grub.peimage,2 in latest
grub,4 is CVE-2023-4692 CVE-2023-4693 ntfs driver fixes
grub.peimage,2 is CVE-2024-2312

The grub.debian,4 is no longer needed, as grub,4 implies
it.

It is not fully clear if additional Debian/Ubuntu downstreams
signed grub with the Canonical peimage loader without including
a grub.peimage SBAT level. The grub.debian and grub.ubuntu SBAT
levels were bumped as well but we'd prefer to revoke grub.peimage
as it was built specifically for that purpose.

Add the previous latest level to the switch for automatic.

Signed-off-by: Julian Andres Klode <julian.klode@canonical.com>